Move some utilities out to different modules
