+++ polars_utils/pipeline_utils.py
@@ -1,0 +1,46 @@
+"""Provides some utilities for timing and logging polars pipelines."""
+
+import logging
+import sys
+import time
+from contextlib import contextmanager
+
+import polars as pl
+
+# ECS/Cloudwatch captures stdout logging.
+logging.basicConfig(
+    level=logging.INFO,
+    format="[%(levelname)s] %(message)s",
+    handlers=[logging.StreamHandler(sys.stdout)],
+)
+logger = logging.getLogger(__name__)
+
+
+def log_polars_plan(lf: pl.LazyFrame, context: str) -> None:
+    """Logs the explain plan and schema to CloudWatch immediately."""
+    logger.info(f"--- PRE-FLIGHT CHECK: {context} ---")
+
+    plan = lf.explain(engine="streaming")
+
+    # We log line-by-line so CloudWatch doesn't truncate a massive single string.
+    for line in plan.split("\n"):
+        if line.strip():  # Skip empty lines
+            logger.info(f"[PLAN] {line}")
+
+    # Log the schema too.
+    logger.info(f"Schema for {context}: {lf.collect_schema()}")
+
+    logger.info(f"--- END PRE-FLIGHT PLAN: {context} ---")
+    sys.stdout.flush()
+
+
+@contextmanager
+def time_it(label: str):
+    """Context manager to time code execution."""
+    start = time.perf_counter()
+    try:
+        yield
+    finally:
+        elapsed = time.perf_counter() - start
+        logger.info(f"[METRIC] {label}: {elapsed:.4f}s")
+        sys.stdout.flush()
